Updating LapCounter to use new PathingTool (WIP)

+++ 3D Mario Kart/Assets/Scripts/Path/Checkpoint.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public class Checkpoint : MonoBehaviour
 {
-    public float checkpointID = 0;
+    public int checkpointID = 0;
 }
